# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Malaika-booking API",
        "description": "",
        "contact": {
            "email": "kirian@fruntera.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/departures": {
            "get": {
                "tags": [
                    "Departures"
                ],
                "summary": "Lista las salidas",
                "description": "Lista las salidas",
                "operationId": "departuresList",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "query",
                        "description": "Id del viaje",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Departure list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DepartureResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departures"
                ],
                "summary": "Crea una nueva salida",
                "description": "Crea una nueva salida",
                "operationId": "createDeparture",
                "requestBody": {
                    "description": "Create departure",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trip_id"
                                ],
                                "properties": {
                                    "trip_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "start": {
                                        "type": "string",
                                        "example": "12/20/2023"
                                    },
                                    "final": {
                                        "type": "string",
                                        "example": "12/28/2023"
                                    },
                                    "price": {
                                        "type": "numeric",
                                        "example": "7800.55"
                                    },
                                    "pax_available": {
                                        "type": "integer",
                                        "example": "20"
                                    },
                                    "individual_supplement": {
                                        "type": "numeric",
                                        "example": "550.00"
                                    },
                                    "state_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "commentary": {
                                        "type": "string",
                                        "example": "Salimos muy temprano"
                                    },
                                    "expedient": {
                                        "type": "integer",
                                        "example": "55688"
                                    },
                                    "taxes": {
                                        "type": "numeric",
                                        "example": "157.65"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Departure created successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/departures/{id}": {
            "put": {
                "tags": [
                    "Departures"
                ],
                "summary": "Actualiza los datos de una salida",
                "description": "Actualiza los datos de una salida",
                "operationId": "updateDeparture",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de departure",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update departure",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "start": {
                                        "type": "string",
                                        "example": "12/20/2023"
                                    },
                                    "final": {
                                        "type": "string",
                                        "example": "12/28/2023"
                                    },
                                    "price": {
                                        "type": "numeric",
                                        "example": "7800.55"
                                    },
                                    "pax_available": {
                                        "type": "integer",
                                        "example": "20"
                                    },
                                    "individual_supplement": {
                                        "type": "numeric",
                                        "example": "550.00"
                                    },
                                    "state_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "commentary": {
                                        "type": "string",
                                        "example": "Salimos muy temprano"
                                    },
                                    "expedient": {
                                        "type": "integer",
                                        "example": "55688"
                                    },
                                    "taxes": {
                                        "type": "numeric",
                                        "example": "157.65"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Departure updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/departures/{id}/rooming": {
            "get": {
                "tags": [
                    "Departures"
                ],
                "summary": "Lista las salidas",
                "description": "Lista las salidas",
                "operationId": "departureRoomingList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del viaje",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Departure list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DepartureResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DepartureResource": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "description": "TODO update data",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Trip title",
                        "type": "string",
                        "example": "Antartida desde el mar"
                    },
                    "description": {
                        "description": "Trip description",
                        "type": "string",
                        "example": "Es un viaje muy chulo"
                    },
                    "commentary": {
                        "description": "Trip commentary",
                        "type": "string",
                        "example": "Especial para ciegos"
                    },
                    "state": {
                        "description": "Trip state",
                        "type": "string",
                        "example": "OPEN"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TripResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}